Place the Jitsi and Plaid Utilities into a Module named for the Utility
