+++ app/utilities/plaid/plaid_utility_policy.rb
@@ -1,4 +1,4 @@
-class PlaidUtility < Utility
+module Plaid
   class PlaidUtilityPolicy < UtilityPolicy
     def permitted_params
       %i[environment secret client_id version]
